+++ lib/flow.rb
@@ -1,32 +1,21 @@
 # frozen_string_literal: true
 
 require "active_model"
-require "active_record"
 require "active_support"
 
-require "spicery"
+require "conjunction"
 require "malfunction"
+require "substance"
 
-require "flow/version"
+require_relative "flow/version"
 
-require "flow/concerns/transaction_wrapper"
+require_relative "flow/errors"
 
-require "flow/malfunction/base"
+require_relative "flow/concerns/transaction_wrapper"
 
-require "flow/flow_base"
-require "flow/operation_base"
-require "flow/state_base"
-require "flow/state_proxy"
+require_relative "flow/malfunction/base"
 
-module Flow
-  class Error < StandardError; end
-
-  class FlowError < Error; end
-  class StateInvalidError < FlowError; end
-
-  class OperationError < Error; end
-  class AlreadyExecutedError < OperationError; end
-
-  class StateError < Error; end
-  class NotValidatedError < StateError; end
-end
+require_relative "flow/flow_base"
+require_relative "flow/operation_base"
+require_relative "flow/state_base"
+require_relative "flow/state_proxy"