+++ lib/flow/flow_base.rb
@@ -10,7 +10,7 @@ require_relative "flow/trigger"
 
 # A **Flow** is a collection of procedurally executed **Operations** sharing a common **State**.
 module Flow
-  class FlowBase < Spicerack::RootObject
+  class FlowBase < Substance::RootObject
     include Conjunction::Junction
     suffixed_with "Flow"