+++ lib/flow/rspec/custom_matchers/have_on_state.rb
@@ -44,7 +44,12 @@ module Flow
 
       def matches?(object)
         @state_expectations.all? do |key, value|
-          expect(object.state.public_send(key)).to match value
+          # If state is actually a StateProxy, we to access the state directly with _state
+          if object.state.respond_to?(:_state, true)
+            expect(object.state.__send__(:_state).public_send(key)).to match value
+          else
+            expect(object.state.public_send(key)).to match value
+          end
         end
       end